Initial pass on GKE worker manager
